# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: API transacional e server-authoritative do IA Aprova.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /integrity/device-bindings:
    post:
      operationId: registerIntegrityDeviceBinding
      tags: [integrity]
      description: Registra de forma idempotente um identificador operacional aleatório de instalação. Não é fingerprint nem prova de aparelho.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterIntegrityDeviceBindingRequest" }
      responses:
        "201":
          description: Vínculo operacional criado ou reproduzido pela mesma chave
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IntegrityDeviceBinding" }
        default: { $ref: "#/components/responses/Problem" }
  /integrity/challenges:
    post:
      operationId: createIntegrityChallenge
      tags: [integrity]
      description: Emite desafio curto, de uso único e vinculado à conta, sessão e instalação. Não é enfileirável offline.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateIntegrityChallengeRequest" }
      responses:
        "201":
          description: Desafio emitido; nonce não persiste em claro no servidor
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IntegrityChallenge" }
        default: { $ref: "#/components/responses/Problem" }
  /me/platform-age-signal/verified:
    post:
      operationId: submitVerifiedPlatformAgeSignal
      tags: [integrity, identity]
      description: Reserva de forma idempotente uma verificação vinculada ao challenge. Na Fase A os adapters Apple/Google falham fechados em indeterminate e esta operação nunca grava server_verified.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VerifiedPlatformAgeSignalEnvelope" }
      responses:
        "200":
          description: Resultado terminal sanitizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IntegrityVerification" }
        "202":
          description: Verificação reservada e ainda protegida por lease
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IntegrityVerification" }
        default: { $ref: "#/components/responses/Problem" }
  /integrity/verifications/{verificationId}:
    get:
      operationId: getIntegrityVerification
      tags: [integrity]
      description: Consulta ownership-scoped; nunca expõe verdict, counter, receipt, nonce, hash ou material de prova.
      parameters:
        - name: verificationId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Resultado terminal sanitizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IntegrityVerification" }
        "202":
          description: Verificação em andamento
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IntegrityVerification" }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: O primeiro request com session token Clerk válido provisiona, de forma idempotente, o UUID interno a partir do subject verificado; nenhuma rota aceita clerk_user_id no corpo.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128 }
...
  responses:
    Problem:
      description: Erro no formato RFC 9457
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
  schemas:
...
    ProblemDetails:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        requestId: { type: string }
        errors:
          type: object
          additionalProperties: { type: array, items: { type: string } }
      required: [type, title, status, requestId]
    AgeBand:
      type: string
      enum: ["under_13", "13_15", "16_17", "18_plus"]
...
    Base64Url32Bytes:
      type: string
      minLength: 43
      maxLength: 43
      pattern: '^[A-Za-z0-9_-]{43}$'
    RegisterIntegrityDeviceBindingRequest:
      type: object
      additionalProperties: false
      properties:
        installationNonce: { $ref: "#/components/schemas/Base64Url32Bytes" }
        platform: { enum: [ios, android] }
      required: [installationNonce, platform]
    IntegrityDeviceBinding:
      type: object
      additionalProperties: false
      properties:
        deviceBindingId: { type: string, format: uuid }
        platform: { enum: [ios, android] }
        environment: { enum: [development, production] }
        status: { enum: [active, revoked] }
        registeredAt: { type: string, format: date-time }
      required: [deviceBindingId, platform, environment, status, registeredAt]
    CreateIntegrityChallengeRequest:
      type: object
      additionalProperties: false
      properties:
        purpose: { const: platform_age_signal }
        platform: { enum: [ios, android] }
        deviceBindingId: { type: string, format: uuid }
        appleKeyId:
          anyOf:
            - $ref: "#/components/schemas/Base64Url32Bytes"
            - type: "null"
      required: [purpose, platform, deviceBindingId, appleKeyId]
    IntegrityChallenge:
      type: object
      additionalProperties: false
      properties:
        challengeId: { type: string, format: uuid }
        nonce: { $ref: "#/components/schemas/Base64Url32Bytes", x-sensitive: true }
        principalBinding: { $ref: "#/components/schemas/Base64Url32Bytes" }
        principalBindingKeyVersion: { type: string, minLength: 1, maxLength: 64, pattern: '^[A-Za-z0-9][A-Za-z0-9._-]*$' }
        canonicalizationVersion: { const: age-integrity-v1 }
        expiresAt: { type: string, format: date-time }
      required: [challengeId, nonce, principalBinding, principalBindingKeyVersion, canonicalizationVersion, expiresAt]
    AgeIntegrityClientContext:
      type: object
      additionalProperties: false
      properties:
        osApiGeneration: { enum: [ios_pre_26, ios_26, ios_27_or_later, android, unknown] }
        ageApiCapability: { enum: [declared_age_range_available, play_age_signals_available, unsupported, unknown] }
        runtimeVersion: { type: string, minLength: 1, maxLength: 120, pattern: '^[A-Za-z0-9][A-Za-z0-9._-]*$' }
        updateLaunch: { enum: [embedded, ota] }
        updateId:
          oneOf:
            - const: embedded
            - type: string
              format: uuid
      required: [osApiGeneration, ageApiCapability, runtimeVersion, updateLaunch, updateId]
    NormalizedPlatformAgeSignal:
      type: object
      additionalProperties: false
      properties:
        platform: { enum: [ios, android] }
        status: { enum: [shared, not_shared, verification_required, not_required, unsupported, error] }
        ageBand:
          anyOf:
            - $ref: "#/components/schemas/AgeBand"
            - type: "null"
      required: [platform, status, ageBand]
    VerifiedPlatformAgeSignalEnvelope:
      oneOf:
        - type: object
          additionalProperties: false
          properties:
            challengeId: { type: string, format: uuid }
            challengeNonce: { $ref: "#/components/schemas/Base64Url32Bytes", x-sensitive: true }
            deviceBindingId: { type: string, format: uuid }
            canonicalizationVersion: { const: age-integrity-v1 }
            clientContext: { $ref: "#/components/schemas/AgeIntegrityClientContext" }
            signal:
              allOf:
                - $ref: "#/components/schemas/NormalizedPlatformAgeSignal"
                - type: object
                  properties: { platform: { const: ios } }
            proof:
              type: object
              additionalProperties: false
              properties:
                provider: { const: apple_app_attest }
                keyId: { $ref: "#/components/schemas/Base64Url32Bytes" }
                assertion: { type: string, minLength: 16, maxLength: 24576, x-sensitive: true }
              required: [provider, keyId, assertion]
          required: [challengeId, challengeNonce, deviceBindingId, canonicalizationVersion, clientContext, signal, proof]
        - type: object
          additionalProperties: false
          properties:
            challengeId: { type: string, format: uuid }
            challengeNonce: { $ref: "#/components/schemas/Base64Url32Bytes", x-sensitive: true }
            deviceBindingId: { type: string, format: uuid }
            canonicalizationVersion: { const: age-integrity-v1 }
            clientContext: { $ref: "#/components/schemas/AgeIntegrityClientContext" }
            signal:
              allOf:
                - $ref: "#/components/schemas/NormalizedPlatformAgeSignal"
                - type: object
                  properties: { platform: { const: android } }
            proof:
              type: object
              additionalProperties: false
              properties:
                provider: { const: google_play_integrity_standard }
                integrityToken: { type: string, minLength: 16, maxLength: 60000, x-sensitive: true }
              required: [provider, integrityToken]
          required: [challengeId, challengeNonce, deviceBindingId, canonicalizationVersion, clientContext, signal, proof]
    IntegrityVerification:
      type: object
      additionalProperties: false
      description: Projeção sanitizada; verdict bruto, counter, receipt, nonce e digests são deliberadamente ausentes.
      properties:
        verificationId: { type: string, format: uuid }
        status: { enum: [verifying, verified, rejected, expired, indeterminate] }
        statusUrl: { type: string, pattern: '^/api/v1/integrity/verifications/[0-9a-f-]{36}$' }
        retryAfterSeconds: { type: [integer, "null"], minimum: 1, maximum: 30 }
        decision: { type: [string, "null"], enum: [verification_unavailable, proof_rejected, challenge_expired, verified, null] }
      required: [verificationId, status, statusUrl, retryAfterSeconds, decision]
...